# server/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MotionMesh API
  version: 1.0.0
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []
paths:
...
  /v1/videos:
    get:
      summary: List videos
      responses:
        '200':
          description: OK
    post:
      summary: Upload video initiation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                size_bytes:
                  type: number
      responses:
        '201':
          description: Created
  /v1/videos/{id}:
    get:
      summary: Get video
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete video
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
...